% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{coefficient_change}
\alias{coefficient_change}
\title{Coefficient Change Between Equations}
\usage{
coefficient_change(eq1, eq2)
}
\description{
Coefficient Change Between Equations
}
\keyword{internal}
