% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{compute_derivative_fd}
\alias{compute_derivative_fd}
\title{Centered Finite Differences}
\usage{
compute_derivative_fd(Z, t = NULL, ...)
}
\arguments{
\item{Z}{Numeric vector of observations.}

\item{t}{Numeric vector of time points.}

\item{...}{Additional arguments (ignored).}
}
\value{
List with derivative vector.
}
\description{
Computes derivatives using centered finite differences.
}
