% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{eval_with_coefs}
\alias{eval_with_coefs}
\title{Evaluate equation with modified coefficients}
\usage{
eval_with_coefs(equation, data, new_coefs)
}
\description{
Evaluate equation with modified coefficients
}
\keyword{internal}
