% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_search.R
\name{find_knee_point}
\alias{find_knee_point}
\title{Find Knee Point in Pareto Front}
\usage{
find_knee_point(x, y)
}
\description{
Uses the maximum curvature method to find the "elbow" of the Pareto front.
}
\keyword{internal}
