% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_search.R
\name{format_equation_string}
\alias{format_equation_string}
\title{Format Equation String}
\usage{
format_equation_string(expression, coefficients)
}
\description{
Format Equation String
}
\keyword{internal}
