% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_search.R
\name{get_pareto_set}
\alias{get_pareto_set}
\title{Get Full Pareto Set}
\usage{
get_pareto_set(results)
}
\arguments{
\item{results}{A symbolic_search_result object.}
}
\value{
List of symbolic_equation objects.
}
\description{
Returns all equations on the Pareto front as a list.
}
