% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_example_data}
\alias{list_example_data}
\title{List Available Example Datasets}
\usage{
list_example_data()
}
\value{
A data.frame with dataset names and descriptions.
}
\description{
Returns information about the example datasets included with the package.
}
\examples{
list_example_data()

}
