% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{plot.bifurcation_analysis}
\alias{plot.bifurcation_analysis}
\title{Plot Bifurcation Diagram}
\usage{
\method{plot}{bifurcation_analysis}(x, ...)
}
\arguments{
\item{x}{Object of class bifurcation_analysis.}

\item{...}{Additional arguments (ignored).}
}
\value{
A ggplot object.
}
\description{
Plot Bifurcation Diagram
}
