% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{plot.cv_result}
\alias{plot.cv_result}
\title{Plot CV Results}
\usage{
\method{plot}{cv_result}(x, type = c("predictions", "folds", "both"), ...)
}
\arguments{
\item{x}{Object of class cv_result.}

\item{type}{Type of plot: "predictions", "folds", or "both".}

\item{...}{Additional arguments (ignored).}
}
\value{
A ggplot object or a list of ggplot objects.
}
\description{
Plot CV Results
}
