% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{plot.trajectory_simulation}
\alias{plot.trajectory_simulation}
\title{Plot Simulated Trajectories}
\usage{
\method{plot}{trajectory_simulation}(
  x,
  observed_data = NULL,
  show_trajectories = TRUE,
  n_show = 20,
  alpha_traj = 0.2,
  ...
)
}
\arguments{
\item{x}{Object of class trajectory_simulation.}

\item{observed_data}{Optional observed data to overlay.}

\item{show_trajectories}{Show individual trajectories?}

\item{n_show}{Number of trajectories to show.}

\item{alpha_traj}{Transparency for trajectories.}

\item{...}{Additional arguments (ignored).}
}
\value{
A ggplot object.
}
\description{
Plot Simulated Trajectories
}
