% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{plot.tvr_derivative}
\alias{plot.tvr_derivative}
\title{Plot Method for TVR Derivative}
\usage{
\method{plot}{tvr_derivative}(x, t = NULL, ...)
}
\arguments{
\item{x}{A tvr_derivative object.}

\item{t}{Time vector (uses attribute if NULL).}

\item{...}{Additional plot arguments.}
}
\value{
Invisibly returns the input object (called for side effects).
}
\description{
Plot Method for TVR Derivative
}
