% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{plot.validation_result}
\alias{plot.validation_result}
\title{Plot Validation Results}
\usage{
\method{plot}{validation_result}(x, ...)
}
\arguments{
\item{x}{Object of class validation_result.}

\item{...}{Additional arguments (ignored).}
}
\value{
A list of ggplot objects (invisible).
}
\description{
Plot Validation Results
}
