% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Time Series Plot}
\usage{
plot_timeseries(
  data,
  var,
  time = NULL,
  show_trend = TRUE,
  highlight_changes = TRUE
)
}
\arguments{
\item{data}{Data frame.}

\item{var}{Variable name to plot.}

\item{time}{Time variable name.}

\item{show_trend}{Add trend line?}

\item{highlight_changes}{Highlight potential structural breaks?}
}
\value{
A ggplot object.
}
\description{
Creates a time series plot with optional trend line and change point detection.
}
