% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{plot_trajectory_2d}
\alias{plot_trajectory_2d}
\title{2D Trajectory Plot}
\usage{
plot_trajectory_2d(
  data,
  x_var,
  y_var,
  time_var = NULL,
  show_arrows = TRUE,
  arrow_spacing = 10,
  show_start_end = TRUE
)
}
\arguments{
\item{data}{Data frame.}

\item{x_var}{First state variable.}

\item{y_var}{Second state variable.}

\item{time_var}{Time variable (for coloring trajectory).}

\item{show_arrows}{Add direction arrows?}

\item{arrow_spacing}{Spacing between arrows (every nth point).}

\item{show_start_end}{Mark start and end points?}
}
\value{
A ggplot object.
}
\description{
Plots the trajectory of a system in the (Z, X) plane, useful for
visualizing attractors and limit cycles.
}
