% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_analysis.R
\name{predict.variance_model}
\alias{predict.variance_model}
\title{Predict from Variance Model}
\usage{
\method{predict}{variance_model}(object, newdata, ...)
}
\arguments{
\item{object}{Variance model object}

\item{newdata}{New data for prediction}

\item{...}{Additional arguments}
}
\value{
Numeric vector of predicted standard deviations.
}
\description{
Predict from Variance Model
}
