% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{print.bifurcation_analysis}
\alias{print.bifurcation_analysis}
\title{Print a Bifurcation Analysis}
\usage{
\method{print}{bifurcation_analysis}(x, ...)
}
\arguments{
\item{x}{Object of class bifurcation_analysis.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Bifurcation Analysis
}
