% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{print.tvr_derivative}
\alias{print.tvr_derivative}
\title{Print Method for TVR Derivative}
\usage{
\method{print}{tvr_derivative}(x, ...)
}
\arguments{
\item{x}{A tvr_derivative object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object (called for side effects).
}
\description{
Print Method for TVR Derivative
}
