% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{print.validation_result}
\alias{print.validation_result}
\title{Print Validation Results}
\usage{
\method{print}{validation_result}(x, ...)
}
\arguments{
\item{x}{Object of class validation_result.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object (called for side effects).
}
\description{
Print Validation Results
}
