% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_search.R
\name{setup_julia_backend}
\alias{setup_julia_backend}
\title{Setup Julia Backend}
\usage{
setup_julia_backend()
}
\value{
Logical indicating if the backend is ready.
}
\description{
Checks if Julia and the required SymbolicRegression.jl package are installed.
}
