% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_search.R
\name{symbolic_search_julia}
\alias{symbolic_search_julia}
\title{Julia Backend for Symbolic Search}
\usage{
symbolic_search_julia(
  target,
  predictors,
  operators,
  constraints,
  n_runs,
  complexity_penalty,
  parsimony_pressure,
  julia_options,
  weights,
  verbose
)
}
\description{
Uses SymbolicRegression.jl for advanced symbolic regression.
}
\keyword{internal}
