% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_search.R
\name{symbolic_search_julia_connector}
\alias{symbolic_search_julia_connector}
\title{Julia Backend via JuliaConnectoR (TCP-based)}
\usage{
symbolic_search_julia_connector(
  target,
  predictors,
  operators,
  constraints,
  n_runs,
  complexity_penalty,
  parsimony_pressure,
  julia_options,
  weights,
  verbose
)
}
\description{
Julia Backend via JuliaConnectoR (TCP-based)
}
\keyword{internal}
