% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_search.R
\name{symbolic_search_juliacall}
\alias{symbolic_search_juliacall}
\title{Julia Backend via JuliaCall (embedding-based, legacy)}
\usage{
symbolic_search_juliacall(
  target,
  predictors,
  operators,
  constraints,
  n_runs,
  complexity_penalty,
  parsimony_pressure,
  julia_options,
  weights,
  verbose
)
}
\description{
Julia Backend via JuliaCall (embedding-based, legacy)
}
\keyword{internal}
