% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{analyze_bifurcations}
\alias{analyze_bifurcations}
\title{Analyze Bifurcations}
\usage{
analyze_bifurcations(
  equation,
  variable,
  parameter,
  param_range = c(-5, 5),
  n_param = 50,
  z_range = c(-10, 10),
  exogenous_values = list()
)
}
\arguments{
\item{equation}{Fitted equation object.}

\item{variable}{Name of the main variable.}

\item{parameter}{Name of the parameter to vary.}

\item{param_range}{Range for parameter values.}

\item{n_param}{Number of parameter values to test.}

\item{z_range}{Range for searching fixed points.}

\item{exogenous_values}{Fixed values for other variables.}
}
\value{
Object of class "bifurcation_analysis".
}
\description{
Examines how fixed points change as a parameter varies.
}
