% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{create_block_folds}
\alias{create_block_folds}
\title{Create Block Folds (for time series)}
\usage{
create_block_folds(n, k, block_size = NULL)
}
\description{
Create Block Folds (for time series)
}
\keyword{internal}
